Add a SConstruct (and a SConscript that can be used by dependent components) to do the iof build
Type scons to build.
This will look in ../* for each component and check them out or
download them if necessary in build/<component>

It isn't really scons....it would take a bit more work to write
some code to do that.   But it builds the components.   After
the first build, it will not rebuild the components unless it
detects a change or a missing library

One can also specify the location of each source tree manually.  For
example, for pmix, one would say PMIX_SRC=<path to pmix source>

'scons install PREFIX=<installation path>' will put all of the
installation components in the directory specified.  If no
directory is specified, it will place them in an install
subdirectory by default.

This isn't quite what I want.   Some of the code in the SConscript
really belongs in some common directory...I'm now looking into
how we might use parts as it has already solved some of the
issues I'm trying to work out.

Change-Id: I6a69c1193f23dade4e5bf7bd175e05d86dbd0482
Reviewed-on: http://review.whamcloud.com/18221
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>