Build the tests using SCons
Added an argument to the components class to specify the
list of libraries a dependent component should include.

Update the build to use scons

Tested build on OSx and Linux.   Had to disable shared
object build for bmi on OSx

Move retrieval of config variable to top level SConstruct
and pass it as an Export.   I was trying to use Configure
in cppr, and it balked being called twice.

Change-Id: I57136920929a0c9e3cc0aeceb57bed470b62705d
Reviewed-on: http://review.whamcloud.com/18303
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>