Ensure that LD_LIBRARY_PATH is correctly formed.
Previously we only appended to LD_LIBRARY_PATH however
if this did not already have a value then the path would begin with
a : which is a signal to glibc to include the current working directory
in the search path.

Also switch from appending to prepending to the path as wolf might have
conflicting OpenMPI libs installed.

Change-Id: Ia6394821a079f29be5aa35c98c281b847a09d72c
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: http://review.whamcloud.com/18988
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>