CORFSHIP-50 Update the iof build process.
Add a install target to install iof.  Make the default install
location install/iof but add a --install-path option to over-ride
this.  Compile the binaries with RPATH set to the install location
so they can be run without setting LD_LIBRARY_PATH.

Modify the test script to add the "bin" directory to PATH and
invoke the programs directly rather than including a full path
for each binary.

Update scons_local to get the correct RPATH usage for deps.

Change-Id: I62df74f3fdf5d3f4b8bf84281e97802e677562ec
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: http://review.whamcloud.com/19278
Reviewed-by: Yulu Jia <yulu.jia@intel.com>
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>