Add version hooks to build script
This allows one to override the latest version of a dependent
component.   This can be used when an API change is made to
one of these components and needs to be fixed.   One just specifies
the minimum build # in the iof build script.  It will always use the
latest between that and the one in the last good iof build

Change-Id: I0b5bfca04b0eaf4afd15be4b0f0e9e1882236b70
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: http://review.whamcloud.com/22035
Tested-by: Jenkins
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>