IOF-183 readdir: Implement readdir RPC.
Add code for readdir().  The CNSS sends a RPC to the IONSS
which then responds with multiple directory entry names
and stat structs.  The CNSS loops over entries, consuming
then as required, until either the buffer is full or a new
RPC is required.

This code is not thread safe in the IONSS (it should use
readdir_r) and does not handle offset at all.

Bulk data is not implemented yet, neither are any advanced
FUSE options for readdir()

Change-Id: Iea92292ba774ecbaab1d50438bbe0a36e36f3991
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/24383
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>