CORFSHIP-337 test: Add test with (cn|io)nss both running locally.
This gives the test code full access to both the export and the
projection so it can make comparisons between them.
Add a sub-test for CORTSHIP-336

Start both the IONSS and CNSS on the local node as part of setup,
then allow multiple test methods to run locally with access to both
the export and import directories.

Change-Id: Ie296795a539ba4b881aaf45e62b193c68057921f
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/24497
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ramya Devi Dhanapa <ramya.d.dhanapal@intel.com>