IOF-370: il Intercept open and close APIs
This puts in place some minimal infrastructure for
building the interception library and interception
a few APIs.   At present, the functions are only forwarded
on to the next entry point.

The library is only built if libioil=[shared|all], with the
default being shared.  Static library is built with static
or all.  Linker script for static compilation is installed
in share/ld-opts.

For dynamic linking, just set LD_PRELOAD
For static linking, add
-Wl,@<path_to_share>/ld-opts/ioil-ld-opts

Change-Id: Ic1459b363161e9fd4a9ea0dcfe1383898459d3ce
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/25080
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>