nose-tests: Rework IOF testing so it can work with nose tests.
Nose uses a more inclusive check to tell what is a test and what
isn't, so rename anything that isn't a test so that it doesn't
have test in the name.

This allows the use of nosetests-3.4 to launch the tests directly
from the source location.

Note that as iof_test_local.py has execute permissions the --exe
option is needed to nosetests to load this.

Change-Id: Ieaac47c2ebc126f2e42a68e442f403c0863292ad
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/24985
Tested-by: Maloo <hpdd-maloo@intel.com>