IOF-374 Fix the multi instance test run failure on HA
Includes a fix to the setup_iof which was failing on Maloo.
The Iofrunner has been updated to use the latest version of
multi instance test runner. Added more debug messages to be
logged for ionss/cnss.
Replaced occurences of IOF_LOG_MASK with CRT_LOG_MASK.
The scons_local changes are pertaining to the latest version
of test runner.

Change-Id: I4f9fa24b3542d6d09af36f445d98faa714eca09c
Signed-off-by: rddhanap <ramya.d.dhanapal@intel.com>
Reviewed-on: https://review.whamcloud.com/25135
Reviewed-by: Martin Mennig <martinx.mennig@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>