IOF-345: debugging.  Improve the way GAHs are saved in debug logs.
Write macros for adding GAHs to logs so that the code doesn't need
to do memory allocation and printf even when debugging is disabled.

This also simplifies the logs considerably, the first time a GAH is
used it's full value is reported, subsequent uses only print enough
to be able to uniqiuely identify it.

Change-Id: I639110cbb1bab2660c65c3d980969bd129e6051c
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/25385
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>