IOF-411 ionss: Add reference tracking to file handles.
Add additional functions to access and refcount file handles on the ionss.
This makes it possible to use file handles safely in a multi-threaded
environment and means they automatically get released after last use.

This does not yet address reusing GAH and file handles for multiple
opens of the same file however is required for that feature.

Change-Id: Ifca31430c7b32e70b0303346f18ca397cebe97d4
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/25575
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>