IOF-385 getattr: Cache inode numbers on handles
Both the CNSS and the IONSS manage data structures for open files,
and normal I/O traffic for open files involves calls to 'fgetattr'.
Upcoming fail-over features require the CNSS know the inode number of
open files, and it would assist to have this in the logs. We store
the inode number for all open files without incurring any additional
RPC or File System calls.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Change-Id: I30ed88370ea67c25c646a557445432bdb6175276
Reviewed-on: https://review.whamcloud.com/25592
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>