IOF-425 il: Add ioctl to get global access handle
Implement a simple ioctl to retrieve the global access handle
for an open file.   The interception library is modified to check
the ioctl on open.   The result is ignored but I've verified that
it correctly retrieves the value.   In a subsequent change, I'll
add fd tracking.

Add a test for the ioctl and a file for some interception library
tests.

Change-Id: Idc363ec719e5cf38046fd189fa4427f5d7f61e69
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/25658
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>