IOF-430 ctrl_fs: Remove /client and make "counter" more generic
CPPR is the only user of /client so we remove it from cnss
visibility.   Also, CPPR would prefer more flexibility.  This change
allows implementation of the same semantics as the counter provides
(see the unit test changes for an example) while also allowing the
implementation to assign arbitrary ids on open.

counter renamed to tracker.

Change-Id: Ie08bcf8595092ec1e7d66e5a0d36a95b2a8e2ff5
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/25710
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Aditi Gaur <aditi.gaur@intel.com>