IOF-437 readdir: Use per-rpc flag to signal end-of-directory.
Move the "last" flag from a per-entry value to a per RPC value.
This means each reply is smaller, and there are fewer replies
per directory.
Correctly handle the last flag on the client, as the last flag
is now returned with a valid directory entry then report the
entry to FUSE before checking the last flag.

Change-Id: I4c0870ea5996a436d1810d39805aacbbc773083e
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/25976
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>