IOF-438: IOF PAM Functional tests on multi instance environment
Moved the single instance tests to the multi instance runner to run
the tests on multiple nodes. The tests are run on the CN. Additionally,
we have a setup to run on the ION before the actual tests which creates
the environment required for certain tests; and a verfication on the ION
which verifies the outcome of the tests.

Fixed the multi instance failure during cleanup; unmounted the
filesystems and removed all the related directories.

Added yml description files for five and two node test clusters
which will be triggered once the respective GA changes are merged.
Another changeset will be later submitted to remove the
iof_multi_instance.yml.

Change-Id: I224240197abf3206a3e2c7e81db8e32a61eac28e
Signed-off-by: Ramya Dhanapal <ramya.d.dhanapal@intel.com>
Reviewed-on: https://review.whamcloud.com/26259
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>