IOF-419: Modified iof_local yml file to run on single node
In the process of splitting the yml files to run on one, two and
five node test clusters, the changes to to local_yml file are
being commited separately.
Modified the host list in the iof_test_local.yml to point to
only one node since the test is a single instance test.

Change-Id: Ia09fdb72055fccd5f9e17e6214282649edb067df
Signed-off-by: Ramya Dhanapal <ramya.d.dhanapal@intel.com>
Reviewed-on: https://review.whamcloud.com/26180
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Martin Mennig <martinx.mennig@intel.com>