IOF-463 il: Handle fcntl
This change handles fcntl.  This is needed for the following commands:
F_DUPFD
F_DUPFD_CLOEXEC
F_SETFL

The dup flags are intercepted to keep track of an IOF file that has
been duplicated in this way.   The set flags command is intercepted
to disable interception for said file.

Changed is_mapped to disabled so we just turn off interception for
things we don't support.

Change-Id: I816706d88bb7c051a6f121f5da9763d616aba945
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/26349
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>