IOF-466 shutdown: Cascade Shutdown RPC to remaining IONSS
For the shutdown sequence, the CNSS sends a "Shutdown" RPC to its
PSR which causes the IONSS to shutdown. This fails when there are
multiple IONSS processes running; only the PSR gets the shutdown RPC
and exits, leaving the remaining IONSS running. The PSR which
receives the shutdown RPC must broadcast it to the remaining IONSS
processes before exiting.

Squashed changeset: 25270

Only exit the IONSS when the last CNSS detaches.

Rather than have the "shutdown" rpc make the ionss shutdown have it
keep track of the number of clients and only exit when the last cnss
process detaches.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Change-Id: I51a9a9caf4c7e5ba57795df49b3d8177fd5161ed
Reviewed-on: https://review.whamcloud.com/26488
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>