IOF-465: tests: Correct test error reporting when command not found.
Handle FileNotFoundError differently so that it can be logged correctly
to avoid incorrect warnings.  It's perfectly possible for the tests
to run without fusermount installed and in this case the error should
be logged and the tests should pass.  With the previous code the test
was passing but there were errors about logging being reported.

Change-Id: I94f35fc5df04aaedf230801b08fba92042b08881
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/26551
Reviewed-by: Martin Mennig <martinx.mennig@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>