IOF-411 ionss: Reuse GAH for multiple file opens.
Reuse GAH for subsequent opens of the same file.  To do this call stat
on the FD after opening the file, then search the open GAH list for
the file (by inode) and verify the flags are identical.  This leads to
a reduction in resources consumed by the IONSS process but does mean an
extra stat call on open.

Further work would be to be less selective about flags to potentially
be more agressive about reusing handles.

Change-Id: I438453df95f7932f6e634381cae79676601814f4
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/26754
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>