CQE-70 fix problem with test set names
test set names in the results files need to include the
node name if executed by Multi Runner. Without this change
when the Maloo results file is created there can be more
than one test set with the same name. also, there is no
way to map the test to log files

Change-Id: Ib71e13c233b94f0787eb44287a288c91ca87207e
Signed-off-by: mmennigx <martinx.mennig@intel.com>
Reviewed-on: https://review.whamcloud.com/26918
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>