IOF-394 ionss: Add command line options for ionss.
Add options for CaRT group name and progress interval.
CaRT group name is in preperation for allowing the CNSS to attach
to multiple IONSS applications.  Progress interval is to allow
benchmarking to measure different values.

Change around the progress thread to take a base pointer
rather than a crt_ctx so that additional options can be
passed in.

Change-Id: I3b5dd3dc6706fd1f1fb96b7119cad7ca70fb699d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/26942
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>