IOF-491 il: Add more data to ioctl
We need to get the projection id from the ioctl.  Also
add a version so we don't try to interpret an old ioctl
protocol if the size of the struct is the same.

Add the cnss_id to ensure we only bypass CNSS we are
using.

Check the version at startup.

Change-Id: I7cdce2937a2ac04dc4f8249f10662da0e3eb6983
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/26957
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>