CART-195 ionss: Do not check the return code of crt_(req|reply)_get().
CaRT guarantees this functions will work or the RPC callback itself
would not have been invoked so remove the checking of these functions
from IONSS.
This cleans up the code and removes a lot of untested (and impossible
to reach) error paths.  Doing this resolves a lot of compile errors
about reading unitialised data.

This commit modified the VALIDATE_ARGS_STR() macro and call sites.

Change-Id: Ifa0bfb601cddf964ba08fbf50396d3fc7d13b9e8
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27062
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>