CQE-84 Maloo results not being reported correctly
we are overwriting the top results file when Multi Runner executes more
than one description file. Each time Test Runner (or Multi Runner)
completes a description file it creates a subtest_results.yml file.
These files get collected at the end of the test job by the CI to create
the Maloo results.yml.

The file is being written to the top directory TestRun for each
description file Multi Runner is processing. It should be writing this
file in the top level directory for the description files. Which is a
directory under testRun.

Change-Id: I08216508a7390956d8db6aa0d51c1a8f672fded1
Signed-off-by: mmennigx <martinx.mennig@intel.com>
Reviewed-on: https://review.whamcloud.com/27072
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>