IOF-531 ionss: Keep a cache of preallocated file handles.
Maintain a free list of pre-allocated file handles in a list
hung off base.  This means that in many cases when opening a file
there will be no allocation on the critical path leading to
higher throughput.

When allocating a file handle check the free list and use a
pre-allocated one if possible.
When closing a file handle put the handle on the free list if
the list is empty.
In functions which allocate a handle call a new function to
pre-allocate a new handle outside of the critical path.

The mainatined list is at most length 1 however this could well
be extended to be a larger number of perhaps a proportion of
the number of in-use handles.

Change-Id: Ia8be149bd0265a451aaf63c881702a142da4c36e
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27129
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>