IOF-521 ctrl: Ctrl fs needs to be removed on cnss startup errors
When the cnss starts up and has a problem, we need to properly
cleanup the ctrl fs to avoid leaving stale mounts around.

This change works.   It's a bit hacky to execute a stat on the
ctrl_prefix to force the shutdown to take effect but touching
the shutdown file is normally what is done but it doesn't exist
at the point where this is necessary so instead, we invoke the
stop routine and a stat to force the fuse thread to actually
exit.

Change-Id: I85e6c9a440c62535ee569392eed0df06b215f812
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/27218
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>