IOF-526 startup: Overhaul cnss/FUSE startup procedure.
Try to be more consistent and reliable when starting up.

Start threads for FUSE filesystems as the filessytem are created
rather than after all plugins initialise.
Create a new file in the control filesystem to signify the CNSS is
ready to be used.  Migrate testing code away from checking for the
presense of the shutdown file to checking the contents of the active
file.
This means that if the CNSS fails to start for any reason the shutdown
file can still be used to stop it, and df and other operations on
FUSE filesystems will not deadlock the local filesystem.

Change-Id: I1d998b87b9683310979517e26869ca902365c4ef
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27245
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>