IOF-527 ctrlfs: Add new uint variable type to ctrl fs.
Add a new file type which supports an integer value natively.
Read and write callbacks which work directly with integer values
are registered with the file.
Reads always work and return the value in the file contents.
Writes are optional and if provided can reject writes, for example
if a invalid value is set.

Updated the shutdown file to use the new API, the value of shutdown
can now be read, shutdown is only triggered if "1" is written to the
file and touching the file no longers triggers shutdown.

Change-Id: Ib76863e3a16cb19fa20ecf355b2b45fa9258561e
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27289
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>