IOF-343 readdir: Make readdir size configurable.
Add a ionss command line option to make the size of the readdir bulk buffers
configurable.  Parse this option on the command line, include the value in
the reply to the query RPC and configure the client to use this size.

If the ionss receives a request for more than this then use the maximum
size instead.

Change-Id: Ibd088c7bbb6daaecc506161bf503d774fc677bc3
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27427
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>