IOF-534 threading: Add option to enable multiple FUSE threads.
Add the option is using multiple fuse threads.

Add a flag to the projection descriptor returned by the signon RPC which
if set causes the CNSS to enable FUSE threads.  This is enabled on a
per-projection basis but it currently hard-coded to ON on the IONSS.

Fix a issue with how iof_fuseops was handling the flags/mode bitmask.

This adds a command line option and associated code however it does not
turn on the feature.

Change-Id: Ib148173172b89126149b06be0799d4d98a4c5fef
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27452
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>