IOF-343 log: Minor change in log to turn ERR into INFO
The default behavior of the the ctrl_fs library is to search
for a suitable ctrl fs.   If multiple users have active
ctrl fs running, error messages will print for every failed
attempt to read from one.   This is confusing when a later
attach succeeds.  Such messages should be informational
instead and leave the error for when attach has failed
completely.

Change-Id: I890c7632d39b3a446c87134c3c5e55b6f77293d2
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/27586
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>