IOF-494 cnss: Associate a fs_handle with a cart endpoint.
When either a fs or a dir handle is created keep a copy of the
cart ep along with it so that any future RPCs relating to that
handle can be addressed to the correct location.
This should be matched to the "root" entry in gah however keeping
a copy like this means it doesn't need to be recomputed each
time.

Change-Id: Ida4bc4648bcc21a883a7caa4ed4ff74006fd2d20
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27643
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>