IOF-542 ctrlfs: Better handling of st_size for ctrl fs.
We were seeing a bug where on files accessed concurrently by
multiple processes sometimes the data read from a file would be
padded to 4k with NULL data.  What appeared to be happening was
that getattr() would return a size of 4k, read() would be called
and return less data than then so then getattr() would be called
again followed by a second read(), presumably as the kernel tried
to test for short reads or reconcile the difference between the
stated file size and the amount read.

Fix this by updating the st_size value in read() so that after
the first read getattr() will return consistent size.

Create a open_handle datastructure to track currently open files
and use this to avoid the need for node lookups for every read/
write of getfattr() call.  For open files track the size for
the fd rather than the file itself in case of different values
(for example tracker).

For files where getattr() is called before open stick to the
current behaviour of using 4k as a file size becuase if zero
is used the read() callback does not get invoked.  If a default
value of 1 is used then short reads are returned to the user
and some calls might not like this.

Improve the handle of constants by reducing the maximum lengh,
not registering the file if the length is exceeded (rather than
truncating) and pre-computing the file length.

This patch introduces a memory leak for the filesystem handle
of the FUSE shutdown file.  Add a memcheck suppression to silence
this warning for now.

Change-Id: Ie3230fb1667cf2b0ec3cb33877dfa04d764efe52
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27684
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>