IOF-577 read: Add configurable for max read size via IOV.
Add a configuration option for maximum read request size to
be handled via IOV rather than bulk.  This setting is parsed
from the IONSS command line and returned to the CNSS at signon.

Add a check in the read handler on the IONSS that the max size
has not be violated.

Change-Id: If4b1c284e98525d39f9c0ca2e83faff13e05e8f0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27713
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>