IOF-557 ionss: Add a new struct to describe in-flight reads.
Rather than using stack variables for bulk reads create a new struct
to contain the information required and use this.  This is the first
step to being able to put incoming bulk read requests in a queue and
service them outside of the initial RPC handler.

Update several more crt_(req|reply)_get() callsites to not check
return code.

Change-Id: I2a74d466f524f6eaa1896a1e507e35b92c4144cb
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/27736
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>