IOF-343 ionss: Remove fs_id from handles.
The ionss_file_handle struct contained both a fs_id
and a projection pointer but only ever initialised or
use the fs_id which was then used as an array index
to get to the projection.  Change this to record the
pointer directly.

Do not store the path or fs_id with directory handles.

Remove fs_id from fs_handle as this is duplicated with the
projection pointer.

Change-Id: I53ef5f66220111aab8f99b22b2065b76dcec06ec
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28103
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>