IOF-592 il: Add bypass status query
Add a query to get the bypass status for a file.  This
API lets user know if a file is an IOF file and if
bypass is enabled on the file descriptor.  If bypass
is disabled, it gives user the reason.

In order to implement this, the following changes were
made:
1. Disabling a file doesn't remove it from tracking
2. To implement #1, needed to add some stream tracking
   routines to interception.
3. Logging is done for all IOF files regardless of
   whether bypass is enabled.    Logging is disabled
   for all non-IOF files.

Change-Id: I3411c134ef76da8fcabedb3538be990203d8ac84
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/28433
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>