IOF-343 ionss: Avoid a double-release on shutdown.
As the "root" inode is entered into the file hash table it is closed
on normal shutdown when the hash table is drained, therefore if it
is also closed expliciatly then it gets freed twice.  This did not
initially get caught by memcheck because it gets put on an internal
free list rather than freed() initially.

Change-Id: Id2565b7522572859b54fcdbd52385ffbdb3570ed
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28881
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>