IOF-613 fuse: Add a test for read-only projections.
Add a test that stat works on read-only projections.
Fix a memory leak for low-level projections.

When tests are comparing a file in /usr vs a projection check
for a file to use in /usr rather than hard-coding a name.  This
ensures the file exists (on docker some files may not) and also
gives us some control over the file size.

Change-Id: I298d5750126d1d0912a9060cc68d75068d86376c
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/28893
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Martin Mennig <martinx.mennig@intel.com>