IOF-613 lookup: Include stat iov in reply to lookup RPC.
This was a bug, if the inode was already open and found in the
hash table d_iov_set() wasn't called so the client would return
an error as it didn't have a copy of the stat information.

No tests were failing because of this, however iof_test_local
was failing when running all tests, presuably becuase in that
case there were multiple, repeat operations on the same
projection.

Change-Id: I535f391b331d39027e14b56eaaa6c341a89b53ba
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29568
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>