IOF-343 build: Fix issue with latest CaRT update
The previous patches introduced a CPPR failure due to
IOF not cleaning up the context it creates when the group
attach fails.  This change ensures that crt_context_destroy
is called when attach fails and the progress thread is
stopped.   It doesn't fix all of the other cleanup issues
in that function.

Change-Id: Icf0a922c8f958897d7c7cbe47a59cb5c4d2e0e76
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/29589
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>