IOF-659 read: Update read to low-level FUSE.
This is currently a wrapper around the blocking high-level function.

Call bulk read for reads, returning error as appropiate.  Do not attempt
iov reads.

There are three fuse functions available to reply with the data and little
guidence on which to pick. fuse_reply_iov() however allocates memory and
appears to be similar to fuse_reply_data() so skip that one, but include
code to invoke either fuse_reply_data() or fuse_reply_buf().  Benchmarking
will be required to pick a best option here.

Change-Id: Ic4032861d392343ddad92016c87f9a165a3ad264
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29675
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>