IOF-659 read: Remove the direct read RPC type and code.
Delete the plain read RPC and instead use the read_bulk RPC
everywhere.
Fuse appears to always round up reads to a full page anyway
so the CNSS code was never being used, and read_bulk will
reply with in-line IOV data is the data is small enough so
the only potential downside to this is the interception library
needing to create a bulk descriptor for small reads which would
not be used - potentially even this could be avoided by checking
the requested length and skipping this step if the size is less
that the direct-iov-reply size.

Change-Id: I2fc77f5f7d1f55a08ac0ecb79ca35df1067ddf0d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29676
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>