IOF-659 read: Update read buffers to not use the iof_pool_consume().
Create buf descriptors in the init/reset functions rather than using
the mempool created descriptor itself.

This means that when read needs to yield a descriptor to FUSE it can
still return the mempool desc to the pool, to be repopulated on init.

Remove all trace of the iof_pool_consume() function.

Change-Id: Id74446e851d5bcdd5f0df43a312a35d7e8213296
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29678
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>