CART-495 testing: Add a fault injection test to cover startup.
Use the new fault injection code to test memory allocation failures
on startup.  Create a new test which simply invokes the cnss command,
with an appropiate fault-config.yaml file.  Loop for the first few
memory allocations, trying what happens when each one of them fails.

Continue until no faults are injected, in which case the code will
fail as it cannot connect to the IONSS.

Change-Id: I36b184b7f73a3c80c37001f124bb13c0c3cde2e1
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/29744
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>