IOF-343 progress: Flush queue in progress thread before exiting
CaRT has a known issue with corpc aggregation where the user callback
is invoked on a node before reply aggregation occurs.  In the case of
shutdown, this can result in shutdown processing before we send an
aggregated reply to the caller causing issues with shutdown.  Without
a solution from cart, the BKM is to call crt_progress until it times
out.

Change-Id: I582f16a12e41430583857173915bd2710520173e
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.whamcloud.com/29753
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>