IOF-705 fuse: Fix ie_close() during shutdown.
During normal shutdown procedure the filesystem is set offline and
then remaining inode handles are closed by the CNSS however the
is_close() funciton was checking the projection status and not
sending the close RPC if the projection was off-line.

Correct ie_close() to send the RPC if the projection is offline with
EACCESS as this indicates a normal shutdown.

Improve the logging in inode.c

Change-Id: I17466a3ece99dfd13e0e4cd06c686de118799158
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29861
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>