IOF-714 failover: Use eviction callback for failover handling
Register an eviction handler with CaRT to execute failover logic. The
handler atomically updates the PSR of the group for which the eviction
occurred; or disables the group if no more PSRs remain.  It then
locates all the projections corresponding to the group; if the group
was previously disabled, it marks them offline. Else, it migrates all
open handles to the new PSR. The PSR update and migration must be
completed before the callbacks for individual failed RPCs are invoked,
so they may be able to correctly redirect the RPCs and also use valid
handles for subsequent FUSE operations.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Change-Id: Ifa6cd750e857e404b6a48f48f94b7aa9e5ff667e
Reviewed-on: https://review.whamcloud.com/29905
Reviewed-by: Ashley Pittman <ashley.m.pittman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>