IOF-705 ionss: Embed fd file name into file handle.
In the ionss embed the /proc/self/fd... string into the fd handle
when the handle is created rather than creating it on demand.
This increases memory consumption slightly however reduces work
on create.

Add and use a fh_reset function to track the reference counts better.

Remove a incorrect call to free() on a error path which could have
corrupted memory.

Change-Id: Iaa5f4c97e084af7c51f46bf57567644f8baf7534
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29931
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>