CART-341 refs: Correct some refcounting problems in the CNSS.
Always ensure that any addref() is called before req_send() to
avoid access violations.

It is possible for the RPC to be entirely completed and the callback
to be executed before the req_send() function returns so it is not
safe to defer any addref calls until after req_send().

Change-Id: If2cd3f6580dd2167b3f6b43800c94822ff563394
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.whamcloud.com/29947
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>